Various fixes.

+++ src/chip8emuhostex.hpp
@@ -110,7 +110,7 @@ public:
     Chip8EmulatorOptions& options() { return _options; }
     IChip8Emulator& chipEmu() { return *_chipEmu; }
     bool isHeadless() const override { return true; }
-    uint8_t getKeyPressed() override { return 0; }
+    int getKeyPressed() override { return 0; }
     bool isKeyDown(uint8_t key) override { return false; }
     const std::array<bool,16>& getKeyStates() const override { static const std::array<bool,16> keys{}; return keys; }
     void updateScreen() override {}
